Create simple neural network
For this exercise, you will create a simple neural network example. We'll use a synthetic dataset to predict whether a customer will make a purchase based on two features:

Website visit duration
Number of pages visited.
This is a binary classification problem, and we'll use a small neural network for quick execution.